fix ksswift
