+++ features/src/commonMain/kotlin/sobaya/app/features/randomDog/grid/RandomDogGridViewModel.kt
@@ -22,29 +22,6 @@ class RandomDogGridViewModel(
         getRandomDogUseCase(LIMIT_DOG_COUNT)
     }.stateIn(viewModelScope, SharingStarted.Eagerly, Result.Loading)
 
-    val swState: StateFlow<SwRandomDogGridState> = randomDog.map {
-        when (it) {
-            is Result.Loading -> {
-                SwRandomDogGridState.initialState()
-            }
-            is Result.Success -> {
-                SwRandomDogGridState(
-                    data = it.data,
-                    error = null
-                )
-            }
-            is Result.Error -> {
-                SwRandomDogGridState(
-                    data = null,
-                    error = it.exception?.message
-                )
-            }
-            else -> {
-                throw IllegalStateException("")
-            }
-        }
-    }.stateIn(viewModelScope, SharingStarted.WhileSubscribed(500), SwRandomDogGridState.initialState())
-
     private val _isDataCreated = MutableStateFlow<String?>(null)
     val isDataCreated: StateFlow<String?> = _isDataCreated
