--- features/src/commonMain/kotlin/sobaya/app/features/randomDog/grid/SwRandomDogGridState.kt
@@ -1,20 +1,0 @@
-package sobaya.app.features.randomDog.grid
-
-import sobaya.app.data.dogApi.response.RandomDogResponse
-
-data class SwRandomDogGridState(
-    val error: String?,
-    val data: RandomDogResponse?
-) {
-    val isSuccess: Boolean
-        get() = error == null && data != null
-
-    val isError: Boolean
-        get() = error != null
-    companion object {
-        fun initialState() = SwRandomDogGridState(
-            error = null,
-            data = null
-        )
-    }
-}
