+++ shared/build.gradle.kts
@@ -46,6 +46,10 @@ kotlin {
             iosX64Main.dependsOn(this)
             iosArm64Main.dependsOn(this)
             iosSimulatorArm64Main.dependsOn(this)
+            dependencies {
+                implementation(project(":data"))
+                implementation(project(":util"))
+            }
         }
         val iosX64Test by getting
         val iosArm64Test by getting
