+++ shared/src/iosMain/kotlin/sobaya/app/sample202209/Helper.kt
@@ -1,9 +1,10 @@
 package sobaya.app.sample202209
 
 import org.koin.core.context.startKoin
-import sobaya.app.features.randomDog.grid.SwRandomDogGridState
+import sobaya.app.data.dogApi.response.RandomDogResponse
 import sobaya.app.sample202209.di.module
 import sobaya.lib.local.di.databaseModule
+import sobaya.app.util.Result
 
 fun initKoin() {
     startKoin {
@@ -11,4 +12,8 @@ fun initKoin() {
     }
 }
 
-val randomDogGridState: SwRandomDogGridState = SwRandomDogGridState.initialState()
+
+val randomDogResponse = RandomDogResponse(listOf(), "")
+val result = Result.Success(randomDogResponse)
+val error = Result.Error(Exception())
+val loading = Result.Loading
